move home map and the switch toggle for it to its own component so it can be client side rendered
